# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  name: string;
  country: string;
  householdSize: number;
  /** vegan, vegetarian, omnivore, meat-heavy */
  dietType: string;
  /**
     * none, electric, hybrid, petrol, diesel
     * @nullable
     */
  carType?: string | null;
  /** renewable, mixed, fossil */
  homeEnergySource: string;
  flightsPerYear: number;
  onboardingComplete: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ProfileInput {
  name: string;
  country: string;
  householdSize: number;
  dietType: string;
  /** @nullable */
  carType?: string | null;
  homeEnergySource: string;
  flightsPerYear: number;
  onboardingComplete?: boolean;
}

export interface ProfileUpdate {
  name?: string;
  country?: string;
  householdSize?: number;
  dietType?: string;
  /** @nullable */
  carType?: string | null;
  homeEnergySource?: string;
  flightsPerYear?: number;
  onboardingComplete?: boolean;
}

export interface EmissionEntry {
  id: number;
  /** transportation, energy, food, consumption */
  category: string;
  subcategory: string;
  description: string;
  /** CO2 equivalent in kilograms */
  co2Kg: number;
  date: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface EmissionInput {
  category: string;
  subcategory: string;
  description: string;
  co2Kg: number;
  date: string;
  /** @nullable */
  notes?: string | null;
}

export interface EmissionUpdate {
  category?: string;
  subcategory?: string;
  description?: string;
  co2Kg?: number;
  date?: string;
  /** @nullable */
  notes?: string | null;
}

export interface DashboardSummary {
  totalCo2Kg: number;
  monthlyAverageCo2Kg: number;
  /** Percentage vs global average (negative = better) */
  comparedToAveragePercent: number;
  topCategory: string;
  /** Current action streak in days */
  streak: number;
  /** Total CO2 saved through completed actions */
  savedCo2Kg: number;
  /** User's eco level: seedling, sapling, tree, forest */
  level: string;
  totalEntries: number;
}

export interface TrendPoint {
  period: string;
  co2Kg: number;
  entryCount: number;
}

export interface CategoryBreakdown {
  category: string;
  co2Kg: number;
  percentage: number;
  entryCount: number;
}

export interface Recommendation {
  id: number;
  category: string;
  title: string;
  description: string;
  estimatedSavingKg: number;
  /** easy, medium, hard */
  difficulty: string;
  /** low, medium, high */
  impact: string;
  dismissed: boolean;
  createdAt: string;
}

export interface CompletedAction {
  id: number;
  title: string;
  category: string;
  co2SavedKg: number;
  completedAt: string;
  /** @nullable */
  notes?: string | null;
}

export interface ActionInput {
  title: string;
  category: string;
  co2SavedKg: number;
  /** @nullable */
  notes?: string | null;
}

export interface StreakInfo {
  currentStreak: number;
  longestStreak: number;
  totalActions: number;
  totalSavedKg: number;
  level: string;
  /** Actions needed for next level */
  nextLevelAt: number;
  badges: string[];
}

export type ListEmissionsParams = {
/**
 * Filter by category
 */
category?: string;
/**
 * Max results
 */
limit?: number;
};

export type GetDashboardTrendsParams = {
/**
 * Aggregation period
 */
period?: GetDashboardTrendsPeriod;
};

export type GetDashboardTrendsPeriod = typeof GetDashboardTrendsPeriod[keyof typeof GetDashboardTrendsPeriod];


export const GetDashboardTrendsPeriod = {
  weekly: 'weekly',
  monthly: 'monthly',
} as const;