# lib/api-zod/src/generated/types/actionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface ActionInput {
  title: string;
  category: string;
  co2SavedKg: number;
  /** @nullable */
  notes?: string | null;
}