# lib/api-zod/src/generated/types/completedAction.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CompletedAction {
  id: number;
  title: string;
  category: string;
  co2SavedKg: number;
  completedAt: string;
  /** @nullable */
  notes?: string | null;
}