# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalCo2Kg: number;
  monthlyAverageCo2Kg: number;
  /** Percentage vs global average (negative = better) */
  comparedToAveragePercent: number;
  topCategory: string;
  /** Current action streak in days */
  streak: number;
  /** Total CO2 saved through completed actions */
  savedCo2Kg: number;
  /** User's eco level: seedling, sapling, tree, forest */
  level: string;
  totalEntries: number;
}