# lib/api-zod/src/generated/types/emissionEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface EmissionEntry {
  id: number;
  /** transportation, energy, food, consumption */
  category: string;
  subcategory: string;
  description: string;
  /** CO2 equivalent in kilograms */
  co2Kg: number;
  date: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}