# lib/api-zod/src/generated/types/emissionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface EmissionUpdate {
  category?: string;
  subcategory?: string;
  description?: string;
  co2Kg?: number;
  date?: string;
  /** @nullable */
  notes?: string | null;
}