# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  name: string;
  country: string;
  householdSize: number;
  /** vegan, vegetarian, omnivore, meat-heavy */
  dietType: string;
  /**
     * none, electric, hybrid, petrol, diesel
     * @nullable
     */
  carType?: string | null;
  /** renewable, mixed, fossil */
  homeEnergySource: string;
  flightsPerYear: number;
  onboardingComplete: boolean;
  createdAt: string;
  updatedAt: string;
}