# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  name?: string;
  country?: string;
  householdSize?: number;
  dietType?: string;
  /** @nullable */
  carType?: string | null;
  homeEnergySource?: string;
  flightsPerYear?: number;
  onboardingComplete?: boolean;
}