# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Recommendation {
  id: number;
  category: string;
  title: string;
  description: string;
  estimatedSavingKg: number;
  /** easy, medium, hard */
  difficulty: string;
  /** low, medium, high */
  impact: string;
  dismissed: boolean;
  createdAt: string;
}