# lib/api-zod/src/generated/types/streakInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface StreakInfo {
  currentStreak: number;
  longestStreak: number;
  totalActions: number;
  totalSavedKg: number;
  level: string;
  /** Actions needed for next level */
  nextLevelAt: number;
  badges: string[];
}