# lib/api-zod/src/generated/types/trendPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface TrendPoint {
  period: string;
  co2Kg: number;
  entryCount: number;
}